+++ tests/test_util.py
@@ -61,13 +61,24 @@ def test_get_authenticated_session_study_area():
 def test_get_authenticated_session_http_error():
     responses.add(responses.GET, util.AUTH_URL, status=401)
 
-    with pytest.raises(AuthenticationError, match=r'^Was not able to authenticate with credentials provided.*'):
+    with pytest.raises(
+        AuthenticationError, match=r'^Was not able to authenticate with username and password provided.*'
+    ):
         util.get_authenticated_session('user', 'pass')
 
     with pytest.raises(
         AuthenticationError, match=r'^Was not able to authenticate with \.netrc file and no credentials provided.*'
     ):
         util.get_authenticated_session(None, None)
+
+
+@responses.activate
+def test_get_authenticated_session_with_token_header():
+    session = util.get_authenticated_session(token='token')
+    assert session.headers['Authorization'] == 'Bearer token'
+
+    session = util.get_authenticated_session(username='user', password='pass', token='token')
+    assert session.headers['Authorization'] == 'Bearer token'
 
 
 @responses.activate