+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [7.6.0]
+<<<<<<< edl-token-auth
 
 ### Added
 * Added support for using an [Earthdata Login Bearer Token](https://urs.earthdata.nasa.gov/documentation/for_users/user_token) when authenticating the `HyP3` object. 
+=======
+### Changed
+* `ARIA-S1-GUNW` jobs now uses `reference_date` and `secondary_date` as input instead of a list of granules
+>>>>>>> develop
 
 ## [7.5.0]